test ent optimization
