+++ pronotepy/test_pronotepy.py
@@ -1,11 +1,7 @@
 import datetime
 import unittest
 
-from inspect import getmembers, isfunction
-from functools import partial
-
 import pronotepy
-from pronotepy import ent
 
 client = pronotepy.Client(
     "https://demo.index-education.net/pronote/eleve.html", "demonstration", "pronotevs"
@@ -283,20 +279,6 @@ class TestClientInfo(unittest.TestCase):
             r"^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$",
             "email did not match regex",
         )
-
-
-class TestENT(unittest.TestCase):
-    functions: list
-
-    @classmethod
-    def setUpClass(cls) -> None:
-        cls.functions = getmembers(
-            ent, lambda x: isfunction(x) or isinstance(x, partial)
-        )
-
-    def test_functions(self) -> None:
-        for func in self.functions:
-            self.assertRaises(pronotepy.ENTLoginError, func[1], "username", "password")
 
 
 if __name__ == "__main__":
